REWORK: fix input param name, test cases
